Synthetic linelist of COVID-19 patients

A synthetic dataset containing a simulated linelist of Swiss COVID-19 patients from February to June 2020. The incidence that can be derived from aggregating this synthetic linelist corresponds to or closely matches the incidence numbers reported for Switzerland by the Federal Office of Public Health.

CH_simulated_linelist

Format

A data frame with 31'950 rows and 2 columns:

confirmation_date

date of case confirmation in YYYY-mm-dd format

symptom_onset_date

If simulated, date of onset of symptoms in YYYY-mm-dd format

Details

In this simulated dataset, each row refers to a particular patient. The 'confirmation_date' column refers to the date at which a positive COVID-19 test is reported. For each patient, with a probability < 1, a date of onset of symptoms was drawn from a probability distribution. Otherwise, the date of onset of symptoms was left as NA.