Delay between date of onset of symptoms of COVID-19 and date of case confirmation in Hong Kong

A dataset compiling the delay (in days) between symptom onset and case report for close to 3'000 individual cases. This dataset is truncated in time latest and spans from January 2020 to July 2020. This data was aggregated from the publicly-available linelist data for the COVID-19 epidemic in Hong Kong. The linelist data is published by the Centre for Health Protection in Hong Kong.

HK_delay_data

Format

A data frame with 2,948 rows and 2 variables:

event_date

date of symptom onset in YYYY-mm-dd format

report_delay

number of days between symptom onset and case report

Source

https://www.chp.gov.hk