Convolve delay inputs.

This function is flexible in the type of delay inputs it can handle. Each delay input can be one of:

a list representing a distribution object

a discretized delay distribution vector

a discretized delay distribution matrix

a dataframe containing empirical delay data

convolve_delay_inputs(
  delay_incubation,
  delay_onset_to_report,
  n_report_time_steps = NULL,
  ...
)

Arguments

delay_incubation
Incubation delay. Flexible format.

delay_onset_to_report
Delay between symptom onset and report. Flexible format.

n_report_time_steps
integer. Length of incidence time series. Use only when providing empirical delay data.

...
Arguments passed on to get_matrix_from_empirical_delay_distr

min_number_cases

integer. Minimal number of cases to build the empirical distribution from. TODO add details

min_number_cases_fraction

numeric. Between 0 and 1. If min_number_cases is not provided (kept to NULL), the number of most-recent cases used to build the instant delay distribution is min_number_cases_fraction times the total number of reported delays.

min_min_number_cases

numeric. Lower bound for number of cases used to build an instant delay distribution.

upper_quantile_threshold

numeric. Between 0 and 1. TODO add details

ref_date

Date. Optional. Date of the first data entry in incidence_data

time_step

string. Time between two consecutive incidence datapoints. "day", "2 days", "week", "year"... (see seq.Date for details)

Value

a discretized delay distribution vector or matrix.

Details

see get_matrix_from_empirical_delay_distr for details on the format expected for the empirical delay data.