Compute bagged mean from bootstrapped replicates

If original_values are included, these values are included in the mean computation along with the bootstrapped_values.

.summarise_bagged_mean(
  bootstrapped_values,
  original_values = NULL,
  value_col,
  bootstrap_id_col,
  index_col
)

Arguments

bootstrapped_values
Bootstrap replicates of the original data. Must be a dataframe in the long format with a timestep index column named index_col, a bootstrap replicate index column named bootstrap_id_col, and a value column named value_col. The index column must not contain any NA value.

original_values
Optional. Values of reference used to construct the uncertainty interval around. Typically, these are estimates obtained on the original data. Must be a dataframe with a timestep index column named index_col and a value column named value_col. The index column must not contain any NA value.

value_col
string. Name of the column containing values.

bootstrap_id_col
string. Name of the column containing bootstrap samples numbering. Id 0 must correspond to values associated to the original data.

index_col
string. Name of the index column. The index tracks which data point in bootstrapped values corresponds to which data point in the original values.

Value

a dataframe containing a time step index column named index_col and a column containing bagged mean values called value_col