Estimate Re from incidence and estimate uncertainty with block-bootstrapping

An estimation of the effective reproductive number through time is made with smooth_deconvolve_estimate on the original incidence data. Then, the same estimation is performed on a number of bootstrap samples built from the original incidence data. The estimate on the original data is output along with confidence interval boundaries built from the distribution of bootstrapped estimates.

get_block_bootstrapped_estimate(
  incidence_vector,
  N_bootstrap_replicates = 100,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  shape_incubation,
  scale_incubation,
  shape_onset_to_report = 0,
  scale_onset_to_report = 0,
  estimation_window = 3,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1,
  ref_date = NULL,
  time_step = "day",
  verbose = FALSE
)

Arguments

incidence_vector
numeric

N_bootstrap_replicates
integer. Number of bootstrap samples.

smoothing_method
string. see smooth_incidence.

deconvolution_method
string. see deconvolve_incidence

estimation_method
string. see estimate_Re

shape_incubation
numeric. Shape parameter of the gamma distribution representing the delay between infection and symptom onset.

scale_incubation
numeric. Scale parameter of the gamma distribution representing the delay between infection and symptom onset.

shape_onset_to_report
numeric. Shape parameter of the gamma distribution representing the delay between symptom onset and observation.

scale_onset_to_report
numeric. Scale parameter of the gamma distribution representing the delay between symptom onset and observation.

estimation_window
integer. Only use if estimation_method = "EpiEstim sliding window". see estimate_Re

mean_serial_interval
numeric. see estimate_Re

std_serial_interval
numeric. see estimate_Re

mean_Re_prior
numeric. Mean of prior distribution on Re

ref_date
Date. Optional. Date of the first data entry in incidence_vector

time_step
string. "day", "2 days", "week", "year"... (see seq.Date for details)

#TODO expand on the output

verbose
boolean. see deconvolve_incidence

Value

tibble. Re estimations along with results from each pipeline step.

Details

A base assumption made here is that the delay between infection can be split into two independent delays (with the second delay being optional). The first waiting time corresponds to the time between the infection event and the onset of symptoms. The second (optional) waiting time corresponds to the delay between the onset of symptoms and the case observation. If this second waiting time is omitted, observation is assumed to have happened at the end of the first waiting time. In latter case, the first waiting time distribution does not need to correspond to an incubation period per se, but it must in any case correspond to the delay between infection and case observation.

#TODO clarify input

Examples

#TODO add examples