Infer timeseries of infection events from incidence data of delayed observations

This function takes as input incidence data of delayed observations of infections events, as well as the probability distribution(s) of the delay(s). It returns an inferred incidence of infection events. This function can account for the observations being dependent on future delayed observations. For instance, if the incidence data represents symptom onset events, usually these events are dependent on a secondary delayed observation: a case confirmation typically, or a hospital admission or any other type of event,

get_infections_from_incidence(
  incidence_data,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  delay_incubation,
  delay_onset_to_report = c(1),
  is_partially_reported_data = FALSE,
  delay_distribution_final_report = NULL,
  output_infection_incidence_only = TRUE,
  ref_date = NULL,
  time_step = "day",
  ...
)

Arguments

incidence_data
numeric vector. Incidence data (TODO improve definition)

smoothing_method
string. Method used to smooth the original incidence data. Available options are:

'LOESS', implemented in .smooth_LOESS

deconvolution_method
string. Method used to infer timings of infection events from the original incidence data (aka deconvolution step). Available options are:

'Richardson-Lucy delay distribution', implemented in .deconvolve_incidence_Richardson_Lucy

delay_incubation
TODO add details

delay_onset_to_report
TODO add details

is_partially_reported_data
boolean TODO add details

delay_distribution_final_report
TODO add details

output_infection_incidence_only
boolean. TODO add details

ref_date
Date. Optional. Date of the first data entry in incidence_data

time_step
string. Time between two consecutive incidence datapoints. "day", "2 days", "week", "year"... (see seq.Date for details)

...
Arguments passed on to .smooth_LOESS, .deconvolve_incidence_Richardson_Lucy, merge_outputs, correct_for_partially_observed_data

data_points_incl

integer. Size of the window used in the LOESS algorithm.

degree

integer. LOESS degree.

delay_distribution

numeric square matrix or vector. TODO refactor to estimateR

threshold_chi_squared

numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.

max_iterations

integer. Maximum threshold for the number of iterations in the R-L algorithm.

verbose

Boolean. Print verbose output?

cutoff_observation_probability

value between 0 and 1. Only datapoints for timesteps that have a probability to be observed higher than cutoff_observation_probability are kept. The few datapoints with a lower probability to be observed are trimmed off the tail of the timeseries.

Value