Logistic Regression Pipeline

Calculates optimized coefficients, the confidence interval, and performance metrics for the logistic regression model. Then it returns all of these values in a list

logistic_regression_pipeline(n, lambda, beta_true, B = 20, alpha = 0.05)

Arguments

n

An integer specifying the number of observations

lambda

The rate parameter for the Poisson distribution

beta_true

A vector of true coefficients

B

The number of bootstraps. Default is 20

alpha

The significance level to obtain for the 1 - alpha confidence intervals for Beta. Default is .05

Value

A list of key logistic regression model findings

beta_hat

The optimized coefficients

confidence_interval

The confidence interval

metrics

The key performance metrics for the logistic regression model as a list

Author

Destanie Pitt, Hayden Billmann, Ashton Wise

Examples

# Example usage:
# Set parameters for logistic regression pipeline
n <- 100
lambda <- 3
beta_true <- c(0.5, -1, 2)
B <- 50
alpha <- 0.05
results <- logistic_regression_pipeline(n, lambda, beta_true, B, alpha)
#> Beta Estimates:
#>            [,1]
#> [1,]  0.1203733
#> [2,] -1.4196705
#> [3,]  4.2684526
#> 
#> Confidence Intervals:
#>            [,1]        [,2]      [,3]
#> 2.5%  -4.797686 -21.7794825  2.006993
#> 97.5% 14.109621  -0.5934808 48.864636
#> 
#> Metrics:
#> $prevalence
#> [1] 0.94
#> 
#> $confusion_matrix
#>          Actual
#> Predicted  0  1
#>         0  5  1
#>         1  1 93
#> 
#> $accuracy
#> [1] 0.98
#> 
#> $sensitivity
#> [1] 0.9893617
#> 
#> $specificity
#> [1] 0.8333333
#> 
#> $false_discovery_rate
#> [1] 0.0106383
#> 
#> $diagnostic_odds_ratio
#> [1] 465
#> 
print(results$beta_hat)
#>            [,1]
#> [1,]  0.1203733
#> [2,] -1.4196705
#> [3,]  4.2684526
print(results$confidence_interval)
#>            [,1]        [,2]      [,3]
#> 2.5%  -4.797686 -21.7794825  2.006993
#> 97.5% 14.109621  -0.5934808 48.864636
print(results$metrics)
#> $prevalence
#> [1] 0.94
#> 
#> $confusion_matrix
#>          Actual
#> Predicted  0  1
#>         0  5  1
#>         1  1 93
#> 
#> $accuracy
#> [1] 0.98
#> 
#> $sensitivity
#> [1] 0.9893617
#> 
#> $specificity
#> [1] 0.8333333
#> 
#> $false_discovery_rate
#> [1] 0.0106383
#> 
#> $diagnostic_odds_ratio
#> [1] 465
#>