Estimate Re from incidence data

This pipe function combines a smoothing step using (to remove noise from the original observations), a deconvolution step (to retrieve infection events from the observed delays), and an Re estimation step wrapping around estimate_R.

smooth_deconvolve_estimate(
  incidence_data,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  delay_incubation,
  delay_onset_to_report = c(1),
  ref_date = NULL,
  time_step = "day",
  output_Re_only = TRUE,
  ...
)

Arguments

incidence_data
numeric vector. Incidence data (TODO improve definition)

smoothing_method
string. Method used to smooth the original incidence data. Available options are:

'LOESS', implemented in .smooth_LOESS

deconvolution_method
string. Method used to infer timings of infection events from the original incidence data (aka deconvolution step). Available options are:

'Richardson-Lucy delay distribution', implemented in .deconvolve_incidence_Richardson_Lucy

estimation_method
string. Method used to estimate reproductive number values through time from the reconstructed infection timings. Available options are:

'EpiEstim sliding window', implemented in .estimate_Re_EpiEstim_sliding_window

delay_incubation
TODO add details

delay_onset_to_report
TODO add details

output_Re_only
boolean. Should the output only contain Re estimates? (as opposed to containing results for each intermediate step)

...
Arguments passed on to .smooth_LOESS, .deconvolve_incidence_Richardson_Lucy, .estimate_Re_EpiEstim_sliding_window

data_points_incl

integer. Size of the window used in the LOESS algorithm.

degree

integer. LOESS degree.

delay_distribution

numeric square matrix or vector. TODO refactor to estimateR

threshold_chi_squared

numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.

max_iterations

integer. Maximum threshold for the number of iterations in the R-L algorithm.

verbose

Boolean. Print verbose output?

minimul_cumul_incidence

Numeric scalar. Minimum number of cumulated infections before starting the Re estimation

estimation_window

Integer scalar. Number of data points over which to assume Re to be constant.

mean_serial_interval

Numeric positive scalar. mean_si for estimate_R

std_serial_interval

Numeric positive scalar. std_si for estimate_R

mean_Re_prior

Numeric positive scalar. mean prior for estimate_R

Value

effective reproductive number estimates through time TODO add details

Details

The smoothing step uses the LOESS method by default. The deconvolution step uses the Richardson-Lucy algorithm by default. The Re estimation uses the Cori method with a sliding window by default.