Summarise the uncertainty obtained from bootstrapping

Summarise the uncertainty obtained from bootstrapping

summarise_uncertainty(
  bootstrapped_estimates,
  original_estimates = NULL,
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  Re_estimate_col = "R_mean",
  bootstrap_id_col = "bootstrap_id",
  index_col = "idx",
  ...
)

Arguments

bootstrapped_estimates
Estimates obtained on a number of bootstrap replicates of the original data. Must be a dataframe in the long format with a timestep index column named index_col, a bootstrap replicate index column named bootstrap_id_col, and an estimate column named Re_estimate_col. The index column must not contain any NA value.

original_estimates
Optional. Estimates obtained on the original data. Must be a dataframe with a timestep index column named index_col and an estimate column named Re_estimate_col. The index column must not contain any NA value.

uncertainty_summary_method
One of these options:

'original estimate - CI from bootstrap estimates'. The confidence interval is built using bootstrapped estimates and centered around the original estimates.

'bagged mean - CI from bootstrap estimates'. The confidence interval is built using bootstrapped estimates and centered around the mean of bootstrapped estimates and original estimates.

Re_estimate_col
string. Name of the column containing Re estimates

bootstrap_id_col
string. Name of the column containing bootstrap samples numbering. Id 0 must correspond to the estimate on the original data.

index_col
string. Index column to keep track of which data point in bootstrapped estimates corresponds to which data point in the original estimates.

...
Arguments passed on to .summarise_CI_bootstrap

central_estimates

Estimates around which the confidence interval is going to be centered. Must be a dataframe with a timestep index column named index_col and an estimate column named Re_estimate_col. The index column must not contain any NA value.

alpha

value between 0 and 1. Confidence level of the confidence interval.

Value

dataframe containing Re estimates (column 'Re_estimate') and confidence interval boundaries, with 4 columns like so:

index_col, the timestep index column

Re_estimate_col, Re estimates

CI_up, the upper limit of the confidence interval

CI_down, the lower limit of the confidence interval