Estimate Re from incidence and estimate uncertainty with block-bootstrapping

An estimation of the effective reproductive number through time is made on the original incidence data. Then, the same estimation is performed on a number of bootstrap samples built from the original incidence data. The estimate on the original data is output along with confidence interval boundaries built from the distribution of bootstrapped estimates.

get_block_bootstrapped_estimate(
  incidence_data,
  N_bootstrap_replicates = 100,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  delay_incubation,
  delay_onset_to_report = c(1),
  ref_date = NULL,
  time_step = "day",
  ...
)

Arguments

incidence_data
numeric vector. Incidence data (TODO improve definition)

N_bootstrap_replicates
integer. Number of bootstrap samples.

smoothing_method
string. Method used to smooth the original incidence data. Available options are:

'LOESS', implemented in .smooth_LOESS

deconvolution_method
string. Method used to infer timings of infection events from the original incidence data (aka deconvolution step). Available options are:

'Richardson-Lucy delay distribution', implemented in .deconvolve_incidence_Richardson_Lucy

estimation_method
string. Method used to estimate reproductive number values through time from the reconstructed infection timings. Available options are:

'EpiEstim sliding window', implemented in .estimate_Re_EpiEstim_sliding_window

uncertainty_summary_method
string. One of the following options:

'NONE' if no summary of bootstrap estimates is required

'original estimate - CI from bootstrap estimates'. The confidence interval is built using bootstrapped estimates and centered around the original estimates.

'bagged mean - CI from bootstrap estimates'. The confidence interval is built using bootstrapped estimates and centered around the mean of bootstrapped estimates and original estimates.

delay_incubation
TODO add details

delay_onset_to_report
TODO add details

ref_date
Date. Optional. Date of the first data entry in incidence_data

time_step
string. Time between two consecutive incidence datapoints. "day", "2 days", "week", "year"... (see seq.Date for details)

...
Arguments passed on to .smooth_LOESS, .deconvolve_incidence_Richardson_Lucy, .estimate_Re_EpiEstim_sliding_window

data_points_incl

integer. Size of the window used in the LOESS algorithm.

degree

integer. LOESS degree.

delay_distribution

numeric square matrix or vector. TODO refactor to estimateR

threshold_chi_squared

numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.

max_iterations

integer. Maximum threshold for the number of iterations in the R-L algorithm.

verbose

Boolean. Print verbose output?

minimum_cumul_incidence

Numeric scalar. Minimum number of cumulated infections before starting the Re estimation

estimation_window

Integer scalar. Number of data points over which to assume Re to be constant.

mean_serial_interval

Numeric positive scalar. mean_si for estimate_R

std_serial_interval

Numeric positive scalar. std_si for estimate_R

mean_Re_prior

Numeric positive scalar. mean prior for estimate_R

Value

tibble. Re estimations along with results from each pipeline step. TODO add details

Details

A base assumption made here is that the delay between infection can be split into two independent delays (with the second delay being optional). The first waiting time corresponds to the time between the infection event and the onset of symptoms. The second (optional) waiting time corresponds to the delay between the onset of symptoms and the case observation. If this second waiting time is omitted, observation is assumed to have happened at the end of the first waiting time. In latter case, the first waiting time distribution does not need to correspond to an incubation period per se, but it must in any case correspond to the delay between infection and case observation.